+++ animals-config-starter/build.gradle
@@ -17,13 +17,16 @@ dependencies {
     implementation group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.16.0'
     implementation group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: '2.16.0'
     implementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.16.0'
-    implementation group: 'com.google.code.gson', name: 'gson', version: '2.10.1'
     testImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: '2.7.4'
 
     implementation 'org.hibernate:hibernate-core:6.1.7.Final'
     implementation group: 'jakarta.persistence', name: 'jakarta.persistence-api', version: '3.1.0'
     implementation group: 'org.postgresql', name: 'postgresql', version: '42.7.3'
     implementation group: 'com.fasterxml.jackson.datatype', name: 'jackson-datatype-jsr310', version: '2.16.0'
+    implementation group: 'org.flywaydb', name: 'flyway-core', version: '8.5.13'
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter', version: '2.7.4'
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-data-jpa', version: '2.7.4'
+    implementation group: 'org.hibernate', name: 'hibernate-entitymanager', version: '5.6.15.Final'
 }
 
 tasks.named('test') {
