+++ animals-config-starter/src/main/java/repository/BreedRepository.java
@@ -1,9 +1,9 @@
 package repository;
 
-import dto.Animal;
+import dto.Breed;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnimalsRepository extends JpaRepository<Animal, Integer> {
+public interface BreedRepository extends JpaRepository<Breed, Integer> {
 }
