+++ animals-config-starter/src/main/java/repository/HabitatRepository.java
@@ -1,9 +1,9 @@
 package repository;
 
-import dto.Animal;
+import dto.Habitat;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnimalsRepository extends JpaRepository<Animal, Integer> {
+public interface HabitatRepository extends JpaRepository<Habitat, Integer> {
 }
