+++ animals-config-starter/src/main/java/repository/ProviderRepository.java
@@ -1,9 +1,9 @@
 package repository;
 
-import dto.Animal;
+import dto.Provider;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnimalsRepository extends JpaRepository<Animal, Integer> {
+public interface ProviderRepository extends JpaRepository<Provider, Integer> {
 }
