+++ app/build.gradle
@@ -11,6 +11,7 @@ plugins {
     id 'org.springframework.boot' version '3.2.2'
     id 'io.spring.dependency-management' version '1.0.11.RELEASE'
     id 'application'
+//    id 'org.flywaydb.flyway' version '10.12.0'
 }
 
 repositories {
@@ -22,7 +23,6 @@ dependencies {
     // Use JUnit Jupiter for testing.
     implementation group: 'org.springframework', name: 'spring-context', version: '6.1.3'
     implementation group: 'org.springframework', name: 'spring-web', version: '6.1.3'
-    implementation group: 'javax.annotation', name: 'javax.annotation-api', version: '1.3.2'
     implementation group: 'org.springframework.boot', name: 'spring-boot-autoconfigure', version: '2.7.4'
     implementation group: 'org.springframework.boot', name: 'spring-boot-starter', version: '2.7.4'
     implementation group: 'org.postgresql', name: 'postgresql', version: '42.7.3'
@@ -32,6 +32,9 @@ dependencies {
     implementation 'org.hibernate:hibernate-core:6.1.7.Final'
     implementation group: 'jakarta.persistence', name: 'jakarta.persistence-api', version: '3.1.0'
     implementation group: 'org.postgresql', name: 'postgresql', version: '42.7.3'
+    implementation group: 'org.flywaydb', name: 'flyway-core', version: '8.5.13'
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter', version: '2.7.4'
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-data-jpa', version: '2.7.4'
 
     testImplementation 'org.junit.jupiter:junit-jupiter:5.9.1'
     testImplementation 'org.mockito:mockito-junit-jupiter:3.12.4'
