+++ app/src/main/resources/db/migration/V1_2_0__insert_habitat.sql
@@ -1,0 +1,6 @@
+INSERT INTO animals.habitat (id_area, area)
+VALUES (1, 'Savannah'),
+       (2, 'Jungle'),
+       (3, 'Desert'),
+       (4, 'Forest'),
+       (5, 'House');
