+++ app/src/main/resources/db/migration/V1_3_0__insert_provider.sql
@@ -1,0 +1,5 @@
+INSERT INTO animals.provider (id_provider, name, phone)
+VALUES (1, 'Wildlife Rescue', '+7-917-458-56-78'),
+       (2, 'Animal Shelter', '+7-951-216-19-57'),
+       (3, 'Zoo', '+7-982-659-26-72'),
+       (4, 'Wildlife Conservation', '+7-901-556-21-05');
