+++ app/src/main/resources/db/migration/V1_4_0__insert_animal_type.sql
@@ -1,0 +1,5 @@
+INSERT INTO animals.animal_type (id_type, type_name, is_wild)
+VALUES (1, 'Bear', true),
+       (2, 'Tiger', true),
+       (3, 'Parrot', false),
+       (4, 'Hamster', false);
