+++ app/src/main/resources/db/migration/V1_4_1__insert_breed.sql
@@ -1,0 +1,5 @@
+INSERT INTO animals.breed(id_breed, breed_name)
+VALUES (1, 'Brown Bear'),
+       (2, 'Amur Tiger'),
+       (3, 'Syrian Hamster'),
+       (4, 'Cockatoo parrot');
