+++ app/src/main/resources/db/migration/V1_5_0__insert_animals_habitats.sql
@@ -1,0 +1,5 @@
+INSERT INTO animals.animals_habitats (id_animal_type, id_area)
+VALUES (1, 1),
+       (2, 2),
+       (3, 3),
+       (4, 4);
