+++ app/src/main/resources/db/migration/V1_6_0__insert_animals_provider.sql
@@ -1,0 +1,5 @@
+INSERT INTO animals.animals_provider (id_animal_type, id_provider)
+VALUES (1, 1),
+       (2, 2),
+       (3, 3),
+       (4, 4);
