--- app/src/main/resources/hibernate.cfg.xml
@@ -1,34 +1,0 @@
-<!DOCTYPE hibernate-configuration PUBLIC
-        "-//Hibernate/Hibernate Configuration DTD 3.0//EN"
-        "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd">
-
-<hibernate-configuration>
-
-    <session-factory>
-
-        <property name="connection.url">jdbc:postgresql://localhost:5432/mts</property>
-        <property name="connection.driver_class">org.postgresql.Driver</property>
-        <property name="connection.username">postgres</property>
-        <property name="connection.password">1234567890</property>
-
-        <property name="connection.pool_size">1</property>
-
-        <property name="show_sql">false</property>
-
-        <property name="current_session_context_class">thread</property>
-
-        <property name="hbm2ddl.auto">create-drop</property>
-
-        <property name="hibernate.dbcp.initialSize">5</property>
-        <property name="hibernate.dbcp.maxTotal">20</property>
-        <property name="hibernate.dbcp.maxIdle">10</property>
-        <property name="hibernate.dbcp.minIdle">5</property>
-        <property name="hibernate.dbcp.maxWaitMillis">-1</property>
-
-        <mapping class="dto.db_objects.AnimalType"/>
-        <mapping class="dto.db_objects.Habitat"/>
-        <mapping class="dto.db_objects.Provider"/>
-        <mapping class="dto.db_objects.Animal"/>
-        <mapping class="dto.db_objects.Breed"/>
-    </session-factory>
-</hibernate-configuration>
